Create component for homesubscribe and change layout.
