+++ Client/blogs/src/pages/Home.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import Hero from "../Components/Hero";
 import Blog from "./Blog"
+import HomeSubscribe from "../Components/HomeSubscribe";
+
 const Home = () => {
   return (
     <div className="w-full dark:bg-black overflow-x-hidden">
@@ -9,6 +11,7 @@ const Home = () => {
         {/* Optionally include Gallery and Subscribe when needed */}
       </div>
         <Blog/>
+        <HomeSubscribe/>
     </div>
   );
 };
